Web: move info guide related components into the shared directory (#53679)
* Move info guide context provider to shared directory

* Move sliding panel/info guide related components to shared directory
