+++ web/packages/shared/components/SlidingSidePanel/InfoGuide/index.ts
@@ -16,4 +16,5 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { SlidingSidePanel } from './SlidingSidePanel';
+export * from './InfoGuideContext';
+export * from './InfoGuide';
