+++ web/packages/teleport/src/Bots/List/Bots.test.tsx
@@ -19,10 +19,10 @@
 import { MemoryRouter } from 'react-router';
 
 import { render, screen, userEvent, waitFor } from 'design/utils/testing';
+import { InfoGuidePanelProvider } from 'shared/components/SlidingSidePanel/InfoGuide';
 
 import { botsApiResponseFixture } from 'teleport/Bots/fixtures';
 import { ContextProvider } from 'teleport/index';
-import { InfoGuidePanelProvider } from 'teleport/Main/InfoGuideContext';
 import {
   allAccessAcl,
   createTeleportContext,
