+++ web/packages/teleport/src/Main/Main.test.tsx
@@ -20,12 +20,12 @@ import { MemoryRouter } from 'react-router';
 
 import { ListThin } from 'design/Icon';
 import { fireEvent, render, screen } from 'design/utils/testing';
+import { InfoGuideButton } from 'shared/components/SlidingSidePanel/InfoGuide/InfoGuide';
 
 import { Context, ContextProvider } from 'teleport';
 import { apps } from 'teleport/Apps/fixtures';
 import { events } from 'teleport/Audit/fixtures';
 import { clusters } from 'teleport/Clusters/fixtures';
-import { InfoGuideButton } from 'teleport/components/SlidingSidePanel/InfoGuideSidePanel/InfoGuideSidePanel';
 import { databases } from 'teleport/Databases/fixtures';
 import { desktops } from 'teleport/Desktops/fixtures';
 import { getOSSFeatures } from 'teleport/features';
