+++ web/packages/teleport/src/Users/Users.test.tsx
@@ -19,9 +19,9 @@
 import { MemoryRouter } from 'react-router';
 
 import { fireEvent, render, screen, userEvent } from 'design/utils/testing';
+import { InfoGuidePanelProvider } from 'shared/components/SlidingSidePanel/InfoGuide';
 
 import { ContextProvider } from 'teleport';
-import { InfoGuidePanelProvider } from 'teleport/Main/InfoGuideContext';
 import { createTeleportContext } from 'teleport/mocks/contexts';
 import { Access } from 'teleport/services/user';
